+++ lib/active_shipping/carriers/fedex.rb
@@ -134,6 +134,11 @@ module ActiveShipping
       'TR' => :transfer
     )
 
+    DEFAULT_LABEL_STOCK_TYPE = 'PAPER_7X4.75'
+
+    # Available return formats for image data when creating labels
+    LABEL_FORMATS = ['DPL', 'EPL2', 'PDF', 'ZPLII', 'PNG'] 
+
     def self.service_name_for_code(service_code)
       SERVICE_TYPES[service_code] || "FedEx #{service_code.titleize.sub(/Fedex /, '')}"
     end
@@ -217,8 +222,8 @@ module ActiveShipping
 
             xml.LabelSpecification do
               xml.LabelFormatType('COMMON2D')
-              xml.ImageType('PNG')
-              xml.LabelStockType('PAPER_7X4.75')
+              xml.ImageType(options[:label_format] || 'PNG')
+              xml.LabelStockType(options[:label_stock_type] || DEFAULT_LABEL_STOCK_TYPE)
             end
 
             xml.RateRequestTypes('ACCOUNT')
@@ -604,6 +609,11 @@ module ActiveShipping
           when '9040'
             raise ActiveShipping::ShipmentNotFound, first_notification.at('Message').text
           else
+            raise ActiveShipping::ResponseContentError, StandardError.new(first_notification.at('Message').text)
+          end
+        elsif first_notification.at('Severity').text == 'FAILURE'
+          case first_notification.at('Code').text
+          when '9045'
             raise ActiveShipping::ResponseContentError, StandardError.new(first_notification.at('Message').text)
           end
         end