+++ lib/active_shipping/rate_estimate.rb
@@ -27,6 +27,10 @@ module ActiveShipping
   #   The code of the shipping service
   #   @return [String]
   #
+  # @!attribute description
+  #   Public description of the shipping service (e.g. "2 days delivery")
+  #   @return [String]
+  #
   # @!attribute shipping_date
   #   The date on which the shipment will be expected. Normally, this means that the
   #   delivery date range can only pe prmoised if the shipment is handed over on or
@@ -62,34 +66,47 @@ module ActiveShipping
   # @!attribute insurance_price
   #   The price of insurance in cents.
   #   @return [Integer]
+  #
+  # @!attribute delivery_category
+  #   The general classification of the delivery method
+  #   @return [String]
+  #
+  # @!attribute shipment_options
+  #   Additional priced options bundled with the given rate estimate with price in cents
+  #   @return [Array<{ code: String, price: Integer }>]
+  #
   class RateEstimate
-    attr_reader :origin, :destination, :package_rates,
-                :carrier, :service_name, :service_code,
+    attr_accessor :origin, :destination, :package_rates,
+                :carrier, :service_name, :service_code, :description,
                 :shipping_date, :delivery_date, :delivery_range,
                 :currency, :negotiated_rate, :insurance_price,
                 :estimate_reference, :expires_at, :pickup_time,
-                :compare_price, :phone_required
+                :compare_price, :phone_required, :delivery_category,
+                :shipment_options
 
     def initialize(origin, destination, carrier, service_name, options = {})
-      @origin, @destination, @carrier, @service_name = origin, destination, carrier, service_name
-      @service_code = options[:service_code]
-      @estimate_reference = options[:estimate_reference]
-      @pickup_time = options[:pickup_time]
-      @expires_at = options[:expires_at]
+      self.origin, self.destination, self.carrier, self.service_name = origin, destination, carrier, service_name
+      self.service_code = options[:service_code]
+      self.description = options[:description]
+      self.estimate_reference = options[:estimate_reference]
+      self.pickup_time = options[:pickup_time]
+      self.expires_at = options[:expires_at]
       if options[:package_rates]
-        @package_rates = options[:package_rates].map { |p| p.update(:rate => Package.cents_from(p[:rate])) }
+        self.package_rates = options[:package_rates].map { |p| p.update(:rate => Package.cents_from(p[:rate])) }
       else
-        @package_rates = Array(options[:packages]).map { |p| {:package => p} }
+        self.package_rates = Array(options[:packages]).map { |p| {:package => p} }
       end
-      @total_price = Package.cents_from(options[:total_price])
-      @negotiated_rate = options[:negotiated_rate] ? Package.cents_from(options[:negotiated_rate]) : nil
-      @compare_price = options[:compare_price] ? Package.cents_from(options[:compare_price]) : nil
-      @phone_required = !!options[:phone_required]
-      @currency = ActiveUtils::CurrencyCode.standardize(options[:currency])
-      @delivery_range = options[:delivery_range] ? options[:delivery_range].map { |date| date_for(date) }.compact : []
-      @shipping_date = date_for(options[:shipping_date])
-      @delivery_date = @delivery_range.last
-      @insurance_price = Package.cents_from(options[:insurance_price])
+      self.total_price = options[:total_price]
+      self.negotiated_rate = options[:negotiated_rate]
+      self.compare_price = options[:compare_price]
+      self.phone_required = options[:phone_required]
+      self.currency = options[:currency]
+      self.delivery_range = options[:delivery_range]
+      self.shipping_date = options[:shipping_date]
+      self.delivery_date = @delivery_range.last
+      self.insurance_price = options[:insurance_price]
+      self.delivery_category = options[:delivery_category]
+      self.shipment_options = options[:shipment_options] || []
     end
 
     # The total price of the shipments in cents.
@@ -123,6 +140,40 @@ module ActiveShipping
     # @return [Integer]
     def package_count
       package_rates.length
+    end
+
+    protected
+
+    def delivery_range=(delivery_range)
+      @delivery_range = delivery_range ? delivery_range.map { |date| date_for(date) }.compact : []
+    end
+
+    def total_price=(total_price)
+      @total_price = Package.cents_from(total_price)
+    end
+
+    def negotiated_rate=(negotiated_rate)
+      @negotiated_rate = negotiated_rate ? Package.cents_from(negotiated_rate) : nil
+    end
+
+    def compare_price=(compare_price)
+      @compare_price = compare_price ? Package.cents_from(compare_price) : nil
+    end
+
+    def currency=(currency)
+      @currency = ActiveUtils::CurrencyCode.standardize(currency)
+    end
+
+    def phone_required=(phone_required)
+      @phone_required = !!phone_required
+    end
+
+    def shipping_date=(shipping_date)
+      @shipping_date = date_for(shipping_date)
+    end
+
+    def insurance_price=(insurance_price)
+      @insurance_price = Package.cents_from(insurance_price)
     end
 
     private