+++ test/unit/carriers/ups_test.rb
@@ -128,6 +128,19 @@ class UPSTest < Minitest::Test
                   "DELIVERED"], response.shipment_events.map(&:name)
   end
 
+  def test_find_tracking_info_should_have_messages_for_shipment_events
+    @carrier.expects(:commit).returns(@tracking_response)
+    response = @carrier.find_tracking_info('1Z5FX0076803466397')
+    assert_equal ["BILLING INFORMATION RECEIVED",
+                  "IMPORT SCAN",
+                  "LOCATION SCAN",
+                  "LOCATION SCAN",
+                  "DEPARTURE SCAN",
+                  "ARRIVAL SCAN",
+                  "OUT FOR DELIVERY",
+                  "DELIVERED"], response.shipment_events.map(&:message)
+  end
+
   def test_find_tracking_info_should_have_correct_type_codes_for_shipment_events
     @carrier.expects(:commit).returns(@tracking_response)
     response = @carrier.find_tracking_info('1Z5FX0076803466397')
@@ -229,17 +242,18 @@ class UPSTest < Minitest::Test
   def test_delivery_range_takes_weekend_into_consideration
     mock_response = xml_fixture('ups/test_real_home_as_residential_destination_response')
     @carrier.expects(:commit).returns(mock_response)
-    Timecop.freeze(DateTime.new(2012, 6, 15))
-    response = @carrier.find_rates( location_fixtures[:beverly_hills],
-                                    location_fixtures[:real_home_as_residential],
-                                    package_fixtures.values_at(:chocolate_stuff))
 
-    date_test = [nil, 3, 2, 1, 1, 1].map do |days|
-      DateTime.now.utc + days + 2 if days
+    Timecop.freeze(DateTime.new(2012, 6, 15)) do
+      response = @carrier.find_rates( location_fixtures[:beverly_hills],
+                                      location_fixtures[:real_home_as_residential],
+                                      package_fixtures.values_at(:chocolate_stuff))
+
+      date_test = [nil, 3, 2, 1, 1, 1].map do |days|
+        DateTime.now.utc + (days + 2).days if days
+      end
+
+      assert_equal date_test, response.rates.map(&:delivery_date)
     end
-    Timecop.return
-
-    assert_equal date_test, response.rates.map(&:delivery_date)
   end
 
   def test_maximum_weight
@@ -352,9 +366,9 @@ class UPSTest < Minitest::Test
                                            :billing_zip => expected_postal_code_number,
                                            :billing_country => expected_country_code)
 
-    assert_equal expected_account_number, response.search('ShipmentConfirmRequest/Shipment/PaymentInformation/BillThirdParty/BillThirdPartyShipper/AccountNumber').text
-    assert_equal expected_postal_code_number, response.search('/ShipmentConfirmRequest/Shipment/PaymentInformation/BillThirdParty/BillThirdPartyShipper/ThirdParty/Address/PostalCode').text
-    assert_equal expected_country_code, response.search('/ShipmentConfirmRequest/Shipment/PaymentInformation/BillThirdParty/BillThirdPartyShipper/ThirdParty/Address/CountryCode').text
+    assert_equal expected_account_number, response.search('ShipmentConfirmRequest/Shipment/ItemizedPaymentInformation/ShipmentCharge/BillThirdParty/BillThirdPartyShipper/AccountNumber').text
+    assert_equal expected_postal_code_number, response.search('/ShipmentConfirmRequest/Shipment/ItemizedPaymentInformation/ShipmentCharge/BillThirdParty/BillThirdPartyShipper/ThirdParty/Address/PostalCode').text
+    assert_equal expected_country_code, response.search('/ShipmentConfirmRequest/Shipment/ItemizedPaymentInformation/ShipmentCharge/BillThirdParty/BillThirdPartyShipper/ThirdParty/Address/CountryCode').text
   end
 
   def test_label_request_negotiated_rates_presence
@@ -364,7 +378,8 @@ class UPSTest < Minitest::Test
                                            package_fixtures.values_at(:chocolate_stuff),
                                            :test => true,
                                            :saturday_delivery => true,
-                                           :origin_account => 'A01B23' # without this option, a negotiated rate will not be requested
+                                           :origin_account => 'A01B23', # without this option, a negotiated rate will not be requested
+                                           :negotiated_rates => true,
                              )
 
     negotiated_rates = response.search '/ShipmentConfirmRequest/Shipment/RateInformation/NegotiatedRatesIndicator'